32: extract declareComponent to separate file
